% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exGWAS.R
\docType{data}
\name{exGWAS}
\alias{exGWAS}
\title{An example file of GWAS summary statistics}
\format{
Data frame with columns
\describe{
\item{id}{SNP ID.}
\item{chr}{chromosome.}
\item{pos}{base-pair position.}
\item{A1, A2}{allele codes.}
\item{p}{p value.}
}
}
\usage{
exGWAS
}
\description{
An example file of GWAS summary statistics
}
\examples{
## Load data
data(exGWAS)
}
\keyword{datasets}
