% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snp_to_region.R
\name{map_snp_to_region}
\alias{map_snp_to_region}
\title{Map SNPs to genomic regions (e.g., genes)}
\usage{
map_snp_to_region(
  info_snp,
  info_region,
  extend_region_start = 20L,
  extend_region_end = 20L,
  only_sets = FALSE
)
}
\arguments{
\item{info_snp}{A data frame with columns: "id", "chr", and "pos".
\itemize{
\item id = an unique identifier (e.g., rs numbers)
\item chr = chromosome
\item pos = base-pair position
}}

\item{info_region}{A data frame with columns: "region.id", "chr", "start",
and "end".
\itemize{
\item region.id = an unique identifier (e.g., gene names)
\item chr = chromosome (should use the same chromosome scheme in \code{info_snp})
\item start = start position of genomic region
\item end =  end position of genomic region
}}

\item{extend_region_start}{A single non-negative integer, allowing for a
certain kb window before the region to be included. Default is 20 (= 20kb).}

\item{extend_region_end}{A single non-negative integer, allowing for a
certain kb window after the region to be included. Default is 20 (= 20kb).}

\item{only_sets}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, sets of SNPs for individual
regions are returned, otherwise both sets and mapping information are
returned. Default is \code{FALSE}.}
}
\value{
A nested list containing following components:
\itemize{
\item sets: a named list where each index represents a separate set of SNPs
\item map: a data frame containing SNP mapping information
}
}
\description{
Map SNPs to genomic regions to extract sets of SNPs for set-based tests. A
typical usage would be to annotate SNPs onto their neighboring genes to
perform gene-based tests.
}
