% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_read_bm.R
\name{read.bed.matrix}
\alias{read.bed.matrix}
\title{Read \code{bed.matrix}}
\description{
\code{read.bed.matrix}: Simply re-exporting \code{gaston::read.bed.matrix()} to
end-users. See \code{\link[gaston:read.bed.matrix]{gaston::read.bed.matrix()}} for more details.
}
\keyword{internal}
