% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_snp_readBed.R
\name{snp_readBed}
\alias{snp_readBed}
\title{Read PLINK files into a "bigSNP"}
\description{
Simply re-exporting \code{bigsnpr::snp_readBed()} to end-users.
See \code{\link[bigsnpr:snp_readBed]{bigsnpr::snp_readBed()}} for more details.
}
\keyword{internal}
