% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ref_attach.R
\name{snp_ref_attach}
\alias{snp_ref_attach}
\title{Attach a "bigSNP" reference set from backing files into R.}
\usage{
snp_ref_attach(rdsfile)
}
\arguments{
\item{rdsfile}{A path of ".rds" which stores the \code{bigSNP} object.}
}
\value{
A \code{bigSNP} object.
}
\description{
This is a wrapper function for \code{bigsnpr::snp_attach()} with additional
validations.
}
\details{
It ensures chromosome codes and base-pair positions are integers. If there
are duplicate SNP IDs in the reference set, it attempts to assign an unique
ID with a format: (chr)\emph{(pos)}(allele1)_(allele2). The updated IDs are saved
to \code{rdsfile}.
}
\seealso{
\code{\link[bigsnpr]{snp_attach}} for more details.
}
