% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsettest-package.R
\docType{package}
\name{snpsettest-package}
\alias{snpsettest}
\alias{snpsettest-package}
\title{snpsettest: A set-based association test using GWAS summary statistics for complex-traits}
\description{
The goal of 'snpsettest' is to provide simple tools to compute
             SNP-set level p-values using GWAS summary statistics. The primary
             application of this package is gene-based tests.
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyunjoo@outlook.com}

}
\keyword{internal}
