% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_sumstats.R
\name{harmonize_sumstats}
\alias{harmonize_sumstats}
\title{Harmonizing GWAS summary to reference data}
\usage{
harmonize_sumstats(
  sumstats,
  bigsnpobj,
  join_by_id_only = FALSE,
  remove_ambiguous_snps = TRUE
)
}
\arguments{
\item{sumstats}{A data frame with columns: "snp.id", "chr", "pos", "a0",
"a1" and "p".
\itemize{
\item snp.id = SNP ID (e.g., rs numbers)
\item chr =  chromosome (must be integer)
\item pos =  base-pair position (must be integer)
\item a0 = the first allele
\item a1 = the second allele
\item p = p value of SNP
}

It could have only two columns "snp.id" and "p" if \code{join_by_id_only = TRUE}.}

\item{bigsnpobj}{A \code{bigSNP} object created using the reference data.}

\item{join_by_id_only}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, SNP matching will be
performed by SNP IDs only. Default is \code{FALSE}.}

\item{remove_ambiguous_snps}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the output
removes ambiguous A/T and G/C SNPs for which the strand is not obvious. If
the GWAS genotype data ifself is used as the reference data, it would be
safe to set \code{FALSE}. Default is \code{TRUE}.}
}
\value{
A data frame with columns: "snp.id", "chr", "pos", "a0", "a1" and
"p".
}
\description{
Finds an intersection of variants between GWAS summary and reference data.
}
\details{
Pre-processing of GWAS summary data is required because the sets of variants
available to a particular GWAS might be poorly matched to the variants in
reference data. SNP matching is performed using chromosome code, base-pair
position, and allele codes, while taking into account possible strand flips
and reverse reference alleles. For matched entries, the SNP IDs in summary
data are replaced with the ones in the reference data.
}
