% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reference_bed.R
\name{read_reference_bed}
\alias{read_reference_bed}
\title{Read in reference PLINK Bed file.}
\usage{
read_reference_bed(
  path,
  check_monomorphic = TRUE,
  check_duplicates = TRUE,
  ...
)
}
\arguments{
\item{path}{path to the .bed file}

\item{check_monomorphic}{check monomorphic SNPs and remove them if exists}

\item{check_duplicates}{check duplicate SNPs by ID and genomic position, and
remove them if exists}

\item{...}{passed to \link[gaston:read.bed.matrix]{gaston::read.bed.matrix}}
}
\description{
Read in reference PLINK Bed file.
}
