% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reference_bed.R
\name{read_reference_bed}
\alias{read_reference_bed}
\title{Read in reference PLINK Bed file.}
\usage{
read_reference_bed(path, ...)
}
\arguments{
\item{path}{A path to the .bed file}

\item{...}{Further arguments used in \link[gaston:read.bed.matrix]{gaston::read.bed.matrix}}
}
\description{
Read in reference PLINK Bed file.
}
