% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpset_test.R
\name{snpset_test}
\alias{snpset_test}
\title{Set-based association tests.}
\usage{
snpset_test(
  hsumstats,
  bigsnpobj,
  snp_sets,
  missing_in_geno = TRUE,
  thr_rs = 0.8
)
}
\arguments{
\item{hsumstats}{A data frame processed by \code{\link{harmonize_sumstats}}.}

\item{bigsnpobj}{A \code{bigSNP} object created from the reference data.}

\item{snp_sets}{A named list where each index represents a separate set of
SNPs.}

\item{missing_in_geno}{\code{TRUE} or \code{FALSE}. It determines if genotypes in the
reference data have missing values. If \code{TRUE}, missing genotypes are
imputed by the mode of each SNP. Default is \code{TRUE}.}

\item{thr_rs}{A LD clumping threshold over the squared correlation between
two SNPs. Default is 0.8.}
}
\value{
A data.table with columns: "set.id", "p", "n.snp", "n.snp.clumped",
"top.snp.id" and "top.snp.p"
\itemize{
\item set.id = a name of SNP set
\item p = a set-based p value
\item n.snp = the number of SNPs in an intersection of set input and the
reference data
\item n.snp.clumped = the number of SNPs in a set after LD clumping
\item top.snp.id = SNP ID with the smallest p-value within a set of SNPs
\item top.snp.p = The smallest p-value within a set of SNPs
}
}
\description{
Perform set-based association tests between multiple sets of SNPs and a
phenotype using GWAS summary statistics.
}
