% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpset_test.R
\name{snpset_test}
\alias{snpset_test}
\title{Set-based association tests}
\usage{
snpset_test(hsumstats, x, snp_sets, method = c("davies", "saddle"))
}
\arguments{
\item{hsumstats}{A data frame processed by \code{\link{harmonize_sumstats}}.}

\item{x}{A \code{bed.matrix} object created from the reference data.}

\item{snp_sets}{A named list where each index represents a separate set of
SNPs.}

\item{method}{A method to compute a set-level p value.}
}
\value{
A data.table with columns: "set.id", "p", "n.snp", "n.snp.clumped",
"top.snp.id" and "top.snp.p"
\itemize{
\item set.id = a name of SNP set
\item p = a set-level p value
\item n.snp = the number of SNPs in an intersection of set input and the
reference data
\item n.snp.clumped = the number of SNPs in a set after LD clumping
\item top.snp.id = SNP ID with the smallest p-value within a set of SNPs
\item top.snp.p = The smallest p-value within a set of SNPs
}
}
\description{
Perform set-based association tests between multiple sets of SNPs and a
phenotype using GWAS summary statistics.
}
\examples{
## Load GWAS summary data
data(exGWAS)

## Load reference genotype data
bfile <- system.file("extdata", "example.bed", package = "snpsettest")
x <- read_reference_bed(path = bfile)

## GWAS harmonization with reference data
hsumstats <- harmonize_sumstats(exGWAS, x)

## Load gene information data
data(gene.curated.GRCh37)

## Map SNPs to genes
snp_sets <- map_snp_to_gene(hsumstats, gene.curated.GRCh37)

## Perform set-based (gene-based) association tests
\dontrun{
snpset_test(hsumstats, x, snp_sets$sets)
}
}
