% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsettest-package.R
\docType{package}
\name{snpsettest-package}
\alias{snpsettest}
\alias{snpsettest-package}
\title{snpsettest: A Set-Based Association Test using GWAS Summary Statistics}
\description{
The goal of 'snpsettest' is to provide simple tools that perform
             set-based association tests (e.g., gene-based association tests)
             using GWAS (genome-wide association study) summary statistics. A
             set-based association test in this package is based on the
             statistical model described in VEGAS (versatile gene-based
             association study), which combines the effects of a set of SNPs
             accounting for linkage disequilibrium between markers. This package
             uses a different approach from the original VEGAS implementation to
             compute set-level p values more efficiently, as described in
             <https://github.com/HimesGroup/snpsettest/wiki/Statistical-test-in-snpsettest>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HimesGroup/snpsettest}
  \item Report bugs at \url{https://github.com/HimesGroup/snpsettest/issues}
}

}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyunjoo@outlook.com}

Authors:
\itemize{
  \item Blanca Himes \email{bhimes@pennmeidicine.upenn.edu}
}

}
\keyword{internal}
